Move autolog impl up to spark.py
